@xianlubird 2019-03-28
@jessesuen I have comments in code line, Pls help review again.   And pls add @sarabala1979 help double check that line removed.  

Thank you