@sarabala1979 2019-03-28
Instead of compressing the entire nodestatus on every pod update, We can do an estimate every pad nodestatus size with existing size to check nodestatus will fit or not. This approach will improve the speed on podReconciliation and also this will allow for successful pods which are fit into size. I have already discussed this with @jessesuen.