@xianlubird 2019-03-29
@jessesuen 
I have discussed with @sarabala1979 and already put `checkAndCompress ` back.
Also I have raised a issue #1299  Waiting for @sarabala1979  to implement.  

Thank you .