@xianlubird 2019-03-29
Maybe in the `podReconciliation` function, we can only check `woc.getSize() >= maxWorkflowSize`, if size larger than `maxWorkflowSize`, we `markNodeErrorClearOuput`.  

For most normal process, we shouldn't Marshal all the node status.  

This will speed up process more .